--- Cprogram/tsmap.dat
@@ -1,35 +1,0 @@
-1, 76.0, -233.0,Nishiyonezawaeki, 1, 2
-2, 204.0, -249.0,Seibushougakkou, 5, 1, 3, 4, 17, 33
-3, 177.0, -85.0,Nitori, 3, 2, 4, 6
-4, 281.0, -185.0,Kibamachi, 3, 2, 3, 5
-5, 458.0, -158.0,Housenji, 3, 4, 6, 18
-6, 439.0, -55.0,Tokumachi, 3, 3, 5, 7
-7, 546.0, -64.0,Chuokoukou, 3, 6, 8, 9
-8, 563.0, -16.0,Sougokouen, 1, 7
-9, 789.0, -110.0,Shiyakusho, 3, 7, 10, 16
-10, 1064.0, -174.0,Hanazawayuubinkyoku, 2, 9, 11
-11, 1129, -143,Hanazawa, 2, 10, 12
-12, 1249, -457,Routeinhotel, 2, 11, 13
-13, 1088, -463,Yonezawaeki, 3, 12, 14, 22
-14, 836, -419,Ohmachi, 3, 13, 15, 22
-15, 727, -396,Ohnuma, 4, 14, 16, 19, 21
-16, 743, -317,Tatemachi, 3, 9, 15, 17
-17, 657, -294,Daiichihotel, 4, 2, 16, 18, 19
-18, 675, -192,Nishinomiyaebisujinja, 2, 5, 17
-19, 640, -374,Koujoshougakkou, 3, 15, 17, 20
-20, 628, -428,Montomachi, 4, 19, 21, 26, 33
-21, 715, -453,Hanazen, 4, 15, 20, 22, 24
-22, 860, -475,Aioicho, 4, 13, 14, 21, 23
-23, 922, -717,Hotelheiseii, 2, 22, 24
-24, 670, -651,Nanbushougakkou, 3, 21, 23, 25
-25, 598, -643,Eneosu, 3, 24, 26, 28
-26, 613, -499,Familymart, 3, 20, 25, 27
-27, 540, -477,Uesugijinja, 1, 26
-28, 593, -712,Jonan, 3, 25, 29, 34
-29, 536, -701,Yamadaiseimon, 3, 28, 30, 31
-30, 530, -742,Yamadai, 1, 29
-31, 413, -672,Dainichuugakkou, 2, 29, 32
-32, 340, -636,Yamazawa, 2, 31, 33
-33, 220, -391,Naoecho, 3, 2, 20, 32
-34, 575, -807, Kotohirajinja, 2, 28, 35
-35, 621, -853, Minamiyonezawaeki, 1, 34
